change: move "navigateto" key generation to separate CodeElementXmlDocKeyProvider class, in preparation for porting the logic of the existing XmlDocKeyProvider
